'Init' feature initial implementation
Added the init() proc to babel.nim as well as a new TActionType and
supporting code in doAction.  The idea is that if you have a project
already created in a directory, this will create a .babel file with
sane defaults.  I also updated the help text describe the 'init'
feature as with the other babel commands.