Fix a regression with a package VCS revision
There was a regression with the setting of the right VCS revision to the
`PackageInfo` object when it is being transfered to a full info object.
The regression was caused by the fact that, when the package is
installed and it is not a linked package, then the VCS revision cannot
be obtained from the `.nimble` file directory which is not under version
control, but the revision written in the package meta data should be
used.

Related to nim-lang/nimble#127