Remove validation of develop file dependencies
The validation of develop file dependencies against the Nimble file is
removed in order to allow not direct (transitive) dependencies to be
added to the develop file. Now, all valid packages are allowed to be
added to the project's develop file and not only packages listed in its
`requires` clause. Version range validation for the present in the
Nimble file packages is also removed. On dependency traversal, the
version of the package coming from the develop file is always preferred
instead of an installed package even if its version does not match the
version from the Nimble file. The version ranges of the develop mode
dependencies are validated on `nimble lock` and `nimble check` commands.

Related to nim-lang/nimble#127